// src/app/shared/interfaces/facturas-pendientes.interface.ts
export interface IntFacturasPendiente {
  NroConexion: string;
  RazonSocial: string;
  Documento: string;
  Direccion: string;
  Periodo: string;
  PeriodoDesc: string;
  Importe: string;
  TipoComprobante: string;
  IdComprobante: string;
  PrimerVto: string;
  Comprobante: string;
  corte: string;
  arancel: string;
  permiteFinanciar: string;
}