// src/app/shared/interfaces/facturas-procesadas.interface.ts
export interface IntFacturasProcesadas {
  NroConexion: string;
  RazonSocial: string;
  Documento: string;
  Direccion: string;
  Periodo: string;
  PeriodoDesc: string;
  Importe: string;
  TipoComprobante: string;
  IdComprobante: string;
  PrimerVto: string;
  Comprobante: string;
  corte: string;
  arancel: string;
  permiteFinanciar: string;
  rehabilitacion:boolean
}